+++ tests/Exception/Reflection/CannotResolveClassNameExceptionTest.php
@@ -6,7 +6,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Meritoo\Common\Test\Exception\Reflection;
+namespace Meritoo\Test\Common\Exception\Reflection;
 
 use Generator;
 use Meritoo\Common\Exception\Reflection\CannotResolveClassNameException;
@@ -38,7 +38,7 @@ class CannotResolveClassNameExceptionTest extends BaseTestCase
     public function testConstructorMessage($source, $forClass, $expectedMessage)
     {
         $exception = CannotResolveClassNameException::create($source, $forClass);
-        static::assertEquals($expectedMessage, $exception->getMessage());
+        static::assertSame($expectedMessage, $exception->getMessage());
     }
 
     /**