+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Common and useful classes, methods, exceptions etc.
 
+# 0.1.6
+
+1. Arrays > refactoring & more tests
+2. ValueObject > Human > represents a human
+3. Tests > use `Meritoo\Test\Common` namespace (instead of `Meritoo\Common\Test`)
+4. Tests > use @dataProvider
+
 # 0.1.5
 
 1. Tests > Date > one more test case